consolidate ss props into single hepler
