+++ ui/pages/standards/index.js
@@ -7,9 +7,7 @@ import {
   Filters,
   Dataset,
 } from '../../components';
-import { list, schema } from '../../helpers/api';
-import { parse } from 'url';
-import { parse as qsParse } from 'qs';
+import { getPageProps } from '../../helpers/getPageProps';
 
 export default function Standards({ data, schemaData }) {
   return (
@@ -32,21 +30,6 @@ export default function Standards({ data, schemaData }) {
   );
 }
 
-const DEFAULT_SORT = {
-  column: 'name',
-  order: 'asc',
-};
-
 export async function getServerSideProps(context) {
-  const { resolvedUrl } = context;
-  const parsed = parse(resolvedUrl);
-  const { selections } = qsParse(parsed.query);
-  const { page, sort = DEFAULT_SORT } = context.query;
-
-  return {
-    props: {
-      data: await list({ page, sort, selections }),
-      schemaData: await schema(),
-    },
-  };
+  return await getPageProps(context);
 }
